Moving iiif v2 classes to own files
