+++ src/Wellcome.Dds/IIIF.Tests/IIIF/Presentation/V2/MetadataValueTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using FluentAssertions;
 using IIIF.Presentation.V2;
+using IIIF.Presentation.V2.Strings;
 using IIIF.Presentation.V3.Strings;
 using Xunit;
