+++ src/Wellcome.Dds/IIIF/Presentation/V2/Annotation/IAnnotationListReference.cs
@@ -1,0 +1,7 @@
+﻿namespace IIIF.Presentation.V2.Annotation
+{
+    /// <summary>
+    /// Marker interface for otherContent
+    /// </summary>
+    public interface IAnnotationListReference{}
+}
