+++ src/Wellcome.Dds/IIIF/Presentation/V2/Annotation/IllustrationAnnotationResource.cs
@@ -1,4 +1,4 @@
-namespace IIIF.Presentation.V2
+namespace IIIF.Presentation.V2.Annotation
 {
     public class IllustrationAnnotationResource : LegacyResourceBase
     {
