+++ src/Wellcome.Dds/IIIF/Presentation/V2/Serialisation/MetaDataValueSerialiser.cs
@@ -1,4 +1,5 @@
 using System;
+using IIIF.Presentation.V2.Strings;
 using Newtonsoft.Json;
 
 namespace IIIF.Presentation.V2.Serialisation
