+++ src/Wellcome.Dds/IIIF/Presentation/V2/Strings/LanguageValue.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace IIIF.Presentation.V2
+namespace IIIF.Presentation.V2.Strings
 {
     public class LanguageValue : ValueObject
     {
